+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - Remove support for deprecated `bs-dependencies`, `bs-dev-dependencies`, and `bsc-flags` configuration options. Use `dependencies`, `dev-dependencies`, and `compiler-flags` instead. https://github.com/rescript-lang/rescript/pull/8196
 - `bsc`: remove legacy `-uncurried` flag. https://github.com/rescript-lang/rescript/pull/8201
 - Remove deprecated cli flags `--dev`, `--create-sourcedirs` and `build -w`. https://github.com/rescript-lang/rescript/pull/8202
+- Do not allow two different package specs with the same suffix to avoid conflicts. https://github.com/rescript-lang/rescript/pull/8214
 
 #### :eyeglasses: Spec Compliance